Moved generate_download to homotopy-web/src/model.rs.
